Move AnimatedContentDelegate to a separate file
