+++ lib/src/widgets/widgets.dart
@@ -1,5 +1,6 @@
 export 'animated_book_list.dart';
 export 'animated_book_widget.dart';
+export 'animated_content_delegate.dart';
 export 'animated_content_widget.dart';
 export 'animated_cover_widget.dart';
 export 'background_blur.dart';
